chore: Move translation keys stores to their own folder
