+++ app/javascript/src/stores/translationKeys.js
@@ -1,0 +1,11 @@
+import { writable, derived } from "svelte/store"
+
+export const translationKeys = writable({})
+export const orderedTranslationKeys = derived(translationKeys, $translationKeys =>
+  Object.keys($translationKeys).sort().reduce((result, key) => {
+    result[key] = $translationKeys[key]
+    return result
+  }, {})
+)
+export const selectedLanguages = writable(["en-US"])
+export const defaultLanguage = writable("en-US")
